chore: split keeper from apps
