+++ packages/bindings/src/mocks/mod.rs
@@ -1,3 +1,4 @@
 //! The test utils to mock the quriers and desmos app
 pub mod mock_apps;
 pub mod mock_queriers;
+mod mock_keeper;
